pexor driver: adjusted to SuSe tumbleweed/kernel 5.8.11.1
git-svn-id: https://subversion.gsi.de/dabc/drivers@5086 bcbf6573-9a26-0410-9ebc-ce4ab7aade96

+++ pexor/user/Tests/pexor/polandtest.cpp
@@ -595,7 +595,7 @@ int main (int argc, char **argv)
 
       if(Debugmode)
         printf ("Polled for %d cycles of %d us, tokbuf=0x%x\n", pollcounter, WAITCYCLE, tokbuf);
-      if (tokbuf <=  0 || pollcounter>=MAXPOLLS)
+      if ((unsigned long)tokbuf <=  0 || pollcounter>=MAXPOLLS)
          {
            printf ("\n\nError in Token Request, pollcounter=%d \n",pollcounter);
            return 1;
